refactor(mcp): extract shared helpers, remove duplication across tools
- New Tools::Helpers module with resolve_network_sym, transaction_to_h,
  input_to_h, output_to_h, utxo_to_h, and error_response
- All 6 tools now delegate to Helpers instead of duplicating ~60 lines
- generate_key aligned to use resolve_network_sym (was resolve_network)
- Error responses consistently use Helpers.error_response
